pyproject.toml
[tool.poetry]
name = "rss-reader"
version = "0.0.1"
description = "Simple Rss Reader"
authors = ["hardcorepict <hardcorepict@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
python-dotenv = "^0.20.0"
celery = "^5.2.7"
redis = "^4.3.4"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
django-templated-email = "^3.0.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.6.0"
pytest = "^7.1.2"
factory-boy = "^3.2.1"
Faker = "^13.15.0"
pytest-django = "^4.5.2"
celery = {extras = ["pytest"], version = "^5.2.7"}
pytest-celery = "^0.0.0"
django-querycount = "^0.7.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
